move class Span to a separate header file
