+++ source/extensions/tracers/opentelemetry/BUILD
@@ -31,6 +31,7 @@ envoy_cc_library(
     ],
     hdrs = [
         "opentelemetry_tracer_impl.h",
+        "span.h",
         "span_context.h",
         "span_context_extractor.h",
         "tracer.h",
